requirements.txt
invoke
mock
parameterized
phonenumbers
psycopg2-binary
twilio
unidecode

Django
django-admin-honeypot
django-filter
django-localflavor
django-session-security
django-two-factor-auth
django-widget-tweaks

django-debug-toolbar
djangorestframework